// src/store/main/system/types.ts
// 定义system自己的数据类型接口
export interface ISystemState {
  usersList: any[]
  usersCount: number
  departmentCount: number
  departmentList: any[]
  roleList: any[]
  roleCount: number
  goodsList: any[]
  goodsCount: number
  categoryList: any[]
  categoryCount: number
  menuList: any[]
  menuCount: number
  storyList: any[]
  storyCount: number
}